requirements.txt
fastapi==0.99.1
uvicorn==0.22.0
python-multipart==0.0.6
pydantic==1.10.12
typing-extensions==4.6.0
jinja2==3.1.2
python-docx==0.8.11
PyPDF2==3.0.1